refactor: recoil.ts로 atom 위치 통일
