+++ src/components/Pagination.tsx
@@ -4,7 +4,7 @@ import { AiOutlineArrowLeft } from 'react-icons/ai';
 import { AiOutlineArrowRight } from 'react-icons/ai';
 import { useRouter, useSearchParams } from 'next/navigation';
 import { getIdByNum } from '@/utils/utils';
-import { interviewIdAtom } from '@/utils/atom';
+import { interviewIdAtom } from '@/store/recoil';
 import { useRecoilValue } from 'recoil';
 
 interface PaginatinProps {
