+++ src/components/chat/ChatRoom.tsx
@@ -4,11 +4,11 @@ import ChatBubble from './ChatBubble';
 import ChatInput from './ChatInput';
 
 import useInput from '@/hooks/useInput';
-import { AnswerType } from '@/types/index';
+import { IAnswerInterview } from '@/types/index';
 import { submitAdditionalAnswer } from '@/api/interview';
 
 interface ChatRoomProps {
-    additionalQuestions: AnswerType | undefined;
+    additionalQuestions: IAnswerInterview | undefined;
     handleAnswerCnt: (answerCnt: number) => void;
 }
 
@@ -39,9 +39,9 @@ export default function ChatRoom({
         let regEx = new RegExp(`additional_question_${questionCnt}`);
 
         if (additionalQuestions) {
-            let questionKey: keyof AnswerType = Object.keys(
+            let questionKey: keyof IAnswerInterview = Object.keys(
                 additionalQuestions,
-            ).find((el) => regEx.test(el)) as keyof AnswerType;
+            ).find((el) => regEx.test(el)) as keyof IAnswerInterview;
 
             let newQuestion: ChatMessage = {
                 type: 'question',
