+++ src/utils/utils.ts
@@ -1,14 +1,14 @@
 import { QUESTION_CATEGORY } from '@/constants/data';
 import { useRecoilValue } from 'recoil';
-import { interviewIdAtom } from './atom';
+import { interviewIdAtom } from '@/store/recoil';
 
 export const categoryKeyToName = (key: string) => {
     return QUESTION_CATEGORY.find((category) => category.key === key)?.name;
 };
 
 export const getIdByNum = (num: number) => {
-    const obj = useRecoilValue(interviewIdAtom);
-    return obj[num];
+    const idArr = useRecoilValue(interviewIdAtom);
+    return idArr[num];
 };
 
 export const intToString = (num: number) => String(num).padStart(2, '0');
